WIP: Removed URL analyzing
